# server/api/handler/types.gen.go
// Package handler provides primitives to interact the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen DO NOT EDIT.
package handler

import (
	"time"
)

// Error defines model for Error.
type Error struct {

	// エラーメッセージ
	Message *string `json:"message,omitempty"`
}

// StockPrice defines model for StockPrice.
type StockPrice struct {

	// 終値
	Close string `json:"close"`

	// 取得時間
	Date time.Time `json:"date"`

	// 高値
	High string `json:"high"`

	// 安値
	Low string `json:"low"`

	// 始値
	Open string `json:"open"`
}

// InMaintenance defines model for InMaintenance.
type InMaintenance struct {
	Code *int `json:"code,omitempty"`

	// エラー時のレスポンス
	Data *Error `json:"data,omitempty"`
}

// Unauthorized defines model for Unauthorized.
type Unauthorized Error

// AP01GetStockParams defines parameters for AP01GetStock.
type AP01GetStockParams struct {

	// symbol
	Symbol *string `json:"symbol,omitempty"`
}

// AP01PostStockJSONBody defines parameters for AP01PostStock.
type AP01PostStockJSONBody struct {

	// 登録したい銘柄
	Symbol string `json:"symbol"`
}

// AP01PostStockJSONRequestBody defines body for AP01PostStock for application/json ContentType.
type AP01PostStockJSONRequestBody AP01PostStockJSONBody